+++ packages/react-together/src/hooks/index.ts
@@ -1,3 +1,4 @@
+import useAllNicknames from './useAllNicknames'
 import useChat from './useChat'
 import useConnectedUsers from './useConnectedUsers'
 import useCreateRandomSession from './useCreateRandomSession'
@@ -7,6 +8,7 @@ import useHoveringUsers from './useHoveringUsers'
 import useJoinUrl from './useJoinUrl'
 import useLeaveSession from './useLeaveSession'
 import useMyId from './useMyId'
+import useNicknames from './useNicknames'
 import useStateTogether from './useStateTogether'
 import useStateTogetherWithPerUserValues from './useStateTogetherWithPerUserValues'
 
@@ -15,6 +17,7 @@ import { useIsJoined } from '@croquet/react'
 const useIsTogether = useIsJoined
 
 export {
+  useAllNicknames,
   useChat,
   useConnectedUsers,
   useCreateRandomSession,
@@ -25,6 +28,7 @@ export {
   useJoinUrl,
   useLeaveSession,
   useMyId,
+  useNicknames,
   useStateTogether,
   useStateTogetherWithPerUserValues
 }