+++ packages/react-together/src/hooks/useAllNicknames.ts
@@ -1,0 +1,6 @@
+import useNicknames from './useNicknames'
+
+export default function useAllNicknames() {
+  const [, , allNicknames] = useNicknames()
+  return allNicknames
+}